+++ app/src/main/java/com/safetyapp/mainapp/SettingsActivity.java
@@ -1,0 +1,22 @@
+package com.safetyapp.mainapp;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.util.Log;
+import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+
+public class SettingsActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        setTitle("Settings Page");
+
+
+    }
+}
