+++ app/src/main/java/com/safetyapp/mainapp/SupportConnectionActivity.java
@@ -14,7 +14,11 @@ import java.io.IOException;
 import java.util.*;
 
 
-public class SupportConnectionActivity extends AppCompatActivity {
+public class SupportConnectionActivity extends BaseActivity {
+    @Override
+    protected int getContentLayoutId() {
+        return R.layout.activity_support_connection;
+    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
