+++ app/src/main/res/layout/activity_base.xml
@@ -1,27 +1,54 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
+    android:background="?attr/colorPrimary"
     android:padding="8dp"
-    >
+    android:gravity="center">
 
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="horizontal">
+    <Button
+        android:id="@+id/settings"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:text="Settings"
+        android:textColor="@android:color/white"
+        android:background="?attr/selectableItemBackgroundBorderless"
+        tools:ignore="HardcodedText" />
+
+    <Button
+        android:id="@+id/logout"
+        android:layout_width="0dp"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:text="Log out"
+        android:textColor="@android:color/white"
+        android:background="?attr/selectableItemBackgroundBorderless"
+        tools:ignore="HardcodedText" />
 
     <Button
         android:id="@+id/exitbtn"
         android:layout_width="45dp"
         android:layout_height="45dp"
         android:layout_gravity="end"
-        android:layout_marginTop="30dp"
-        android:layout_marginRight="18dp"
+        android:layout_marginTop="0dp"
+        android:layout_marginRight="0dp"
         android:background="@drawable/btn_circle_red"
         android:text="Exit"
         android:textSize="10dp"
         app:backgroundTint="@null" />
-
+    </LinearLayout>
     <FrameLayout
         android:id="@+id/base_content"
         android:layout_width="match_parent"
+        android:background="@color/white"
         android:layout_height="match_parent"/>
 
 </LinearLayout>
