--- app/src/main/res/layout/fragment_topmenubar.xml
@@ -1,41 +1,0 @@
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="horizontal"
-    android:background="?attr/colorPrimary"
-    android:padding="8dp"
-    android:gravity="center">
-
-    <Button
-        android:id="@+id/settings"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-        android:text="Settings"
-        android:textColor="@android:color/white"
-        android:background="?attr/selectableItemBackgroundBorderless"
-        tools:ignore="HardcodedText" />
-
-    <Button
-        android:id="@+id/logout"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-        android:text="Log out"
-        android:textColor="@android:color/white"
-        android:background="?attr/selectableItemBackgroundBorderless"
-        tools:ignore="HardcodedText" />
-
-    <Button
-        android:id="@+id/exitbtn"
-        android:layout_width="0dp"
-        android:layout_height="wrap_content"
-        android:layout_weight="1"
-        android:text="Emergency"
-        android:textColor="@android:color/white"
-        android:background="?attr/selectableItemBackgroundBorderless"
-        tools:ignore="HardcodedText" />
-
-</LinearLayout>
